Replace netcoreapp1.0 target with netcoreapp2.0
Merged partial classes as all targets now share the same code-base
