+++ Testing/dotNetRDF.MockServerTests/MockRemoteUpdateEndpointFixture.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Net;
+using System.Web;
+using WireMock.RequestBuilders;
+using WireMock.ResponseBuilders;
+using WireMock.Server;
+
+namespace dotNetRDF.MockServerTests
+{
+    public class MockRemoteUpdateEndpointFixture : IDisposable
+    {
+        public readonly FluentMockServer Server;
+
+        public MockRemoteUpdateEndpointFixture()
+        {
+            Server = FluentMockServer.Start();
+            Server.Given(Request.Create()
+                    .WithPath("/update")
+                    .UsingPost()
+                    .WithBody(body=>HttpUtility.UrlDecode(body).StartsWith("update=LOAD <http://dbpedia.org/resource/Ilkeston>")))
+                .RespondWith(Response.Create().WithStatusCode(HttpStatusCode.OK));
+        }
+
+        public void Dispose()
+        {
+            Server.Stop();
+        }
+
+        
+    }
+}
